backend/requirements.txt
# FastAPI core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Async task queue
celery==5.3.4
redis==5.0.1

# WebSocket
websockets==12.0

# LLM API
openai==1.3.7
httpx==0.25.2

# File processing
python-docx==1.1.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
